Test function

+++ app/controllers/lincoln_riders/follow_relationships_controller.rb
@@ -1,18 +1,24 @@
 class LincolnRiders::FollowRelationshipsController < ApplicationController
 	before_action :authenticate_user!
   def create
-	following = current_user.active_follow_relationships.new(followed_user_id: params[:user_id])
-	if following.save
-		redirect_to lincoln_riders_user_mypage_posts_path
+  	if FollowRelationship.find_by(followed_user_id: params[:user_id], following_user_id: current_user.id).nil?
+		following = current_user.active_follow_relationships.new(followed_user_id: params[:user_id])
+		if following.save
+			redirect_to lincoln_riders_user_path(params[:user_id])
+		end
 	else
+		redirect_to lincoln_riders_user_path(params[:user_id])
 	end
   end
 
   def destroy
-  	following = current_user.active_follow_relationships.find_by(followed_user_id: params[:user_id])
-	if following.destroy
-		redirect_to lincoln_riders_user_mypage_posts_path
+  	if !(FollowRelationship.find_by(followed_user_id: params[:user_id], following_user_id:current_user.id).nil?)
+	  	following = current_user.active_follow_relationships.find_by(followed_user_id: params[:user_id])
+		if following.destroy
+			redirect_to lincoln_riders_user_path(params[:user_id])
+		end
 	else
+		redirect_to lincoln_riders_user_path(params[:user_id])
 	end
   end
 end
